requirements.txt
# Base
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt
drf-yasg==1.21.7
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0

# Tools
celery==5.3.6
redis==5.0.1
django-redis==5.4.0


# Debug
django-debug-toolbar==4.2.0
django-silk==5.0.4
Faker==20.0.3
ipython==8.17.2

